rule_based_condition_engine/requirements.txt
pyyaml>=5.4.1
polars>=0.19.0
pyspark>=3.3.0
pyarrow>=10.0.0  # For Parquet support
delta-spark>=2.0.0  # Delta Lake support for PySpark
logging>=0.5.1.2
